requirements.txt
CommonMark==0.7.2
coverage==4.2
future==0.16.0
Jinja2==2.8
MarkupSafe==0.23
mecab-python3==0.7
